# contracts/openapi/powerflow.openapi.json
{
  "components": {
    "schemas": {
...
      "BessMeasurement": {
        "description": "Generator convention: P > 0 discharging.",
        "properties": {
          "alarm_flag_names": {
            "items": {
              "type": "string"
            },
            "title": "Alarm Flag Names",
            "type": "array"
          },
          "alarm_flags": {
            "title": "Alarm Flags",
            "type": "integer"
          },
          "aux_p_kw": {
            "title": "Aux P Kw",
            "type": "number"
          },
          "energy_available_charge_kwh": {
            "title": "Energy Available Charge Kwh",
            "type": "number"
          },
          "energy_available_discharge_kwh": {
            "title": "Energy Available Discharge Kwh",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "limit_flag_names": {
            "items": {
              "type": "string"
            },
            "title": "Limit Flag Names",
            "type": "array"
          },
          "limit_flags": {
            "title": "Limit Flags",
            "type": "integer"
          },
          "operating_state": {
            "title": "Operating State",
            "type": "integer"
          },
          "operating_state_name": {
            "title": "Operating State Name",
            "type": "string"
          },
          "p_available_charge_kw": {
            "title": "P Available Charge Kw",
            "type": "number"
          },
          "p_available_discharge_kw": {
            "title": "P Available Discharge Kw",
            "type": "number"
          },
          "p_cmd_kw": {
            "title": "P Cmd Kw",
            "type": "number"
          },
          "p_kw": {
            "title": "P Kw",
            "type": "number"
          },
          "q_cmd_kvar": {
            "title": "Q Cmd Kvar",
            "type": "number"
          },
          "q_kvar": {
            "title": "Q Kvar",
            "type": "number"
          },
          "s_kva": {
            "title": "S Kva",
            "type": "number"
          },
          "soc_pct": {
            "title": "Soc Pct",
            "type": "number"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "status_name": {
            "title": "Status Name",
            "type": "string"
          },
          "v_lv_kv": {
            "title": "V Lv Kv",
            "type": "number"
          },
          "v_lv_pu": {
            "title": "V Lv Pu",
            "type": "number"
          }
        },
        "required": [
          "id",
          "p_cmd_kw",
          "q_cmd_kvar",
          "p_kw",
          "q_kvar",
          "s_kva",
          "soc_pct",
          "energy_available_discharge_kwh",
          "energy_available_charge_kwh",
          "p_available_discharge_kw",
          "p_available_charge_kw",
          "status",
          "status_name",
          "limit_flags",
          "limit_flag_names",
          "aux_p_kw",
          "v_lv_pu",
          "v_lv_kv",
          "operating_state",
          "operating_state_name",
          "alarm_flags",
          "alarm_flag_names"
        ],
        "title": "BessMeasurement",
        "type": "object"
      },
...
      "BusMeasurement": {
        "properties": {
          "angle_deg": {
            "title": "Angle Deg",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "v_kv": {
            "description": "Line-to-line voltage (kV).",
            "title": "V Kv",
            "type": "number"
          },
          "v_pu": {
            "title": "V Pu",
            "type": "number"
          },
          "vn_kv": {
            "title": "Vn Kv",
            "type": "number"
          }
        },
        "required": [
          "name",
          "vn_kv",
          "v_kv",
          "v_pu",
          "angle_deg"
        ],
        "title": "BusMeasurement",
        "type": "object"
      },
...
      "EngineStatus": {
        "properties": {
          "history_count": {
            "title": "History Count",
            "type": "integer"
          },
          "history_size": {
            "title": "History Size",
            "type": "integer"
          },
          "last_converged": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Converged"
          },
          "last_step_duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Step Duration Ms"
          },
          "nonconverged_count": {
            "title": "Nonconverged Count",
            "type": "integer"
          },
          "overrun_count": {
            "title": "Overrun Count",
            "type": "integer"
          },
          "sim_time": {
            "format": "date-time",
            "title": "Sim Time",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/RunState"
          },
          "step_id": {
            "title": "Step Id",
            "type": "integer"
          },
          "step_s": {
            "title": "Step S",
            "type": "number"
          },
          "test_mode": {
            "title": "Test Mode",
            "type": "boolean"
          }
        },
        "required": [
          "state",
          "test_mode",
          "step_s",
          "step_id",
          "sim_time",
          "last_converged",
          "overrun_count",
          "nonconverged_count",
          "last_step_duration_ms",
          "history_count",
          "history_size"
        ],
        "title": "EngineStatus",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HistoryFormat": {
        "enum": [
          "json",
          "csv"
        ],
        "title": "HistoryFormat",
        "type": "string"
      },
      "HistoryResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "fields": {
            "items": {
              "type": "string"
            },
            "title": "Fields",
            "type": "array"
          },
          "rows": {
            "description": "One row per snapshot: sim_time, step_id and the requested fields.",
            "items": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "integer"
                  },
                  {
                    "type": "string"
                  }
                ]
              },
              "type": "object"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "count",
          "fields",
          "rows"
        ],
        "title": "HistoryResponse",
        "type": "object"
      },
...
      "LoadMeasurement": {
        "description": "Load convention: P > 0 consuming.",
        "properties": {
          "alarm_flag_names": {
            "items": {
              "type": "string"
            },
            "title": "Alarm Flag Names",
            "type": "array"
          },
          "alarm_flags": {
            "title": "Alarm Flags",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "p_kw": {
            "title": "P Kw",
            "type": "number"
          },
          "pf": {
            "title": "Pf",
            "type": "number"
          },
          "q_kvar": {
            "title": "Q Kvar",
            "type": "number"
          },
          "s_kva": {
            "title": "S Kva",
            "type": "number"
          },
          "supply_state": {
            "title": "Supply State",
            "type": "integer"
          },
          "supply_state_name": {
            "title": "Supply State Name",
            "type": "string"
          },
          "v_pu": {
            "title": "V Pu",
            "type": "number"
          }
        },
        "required": [
          "id",
          "p_kw",
          "q_kvar",
          "s_kva",
          "pf",
          "v_pu",
          "supply_state",
          "supply_state_name",
          "alarm_flags",
          "alarm_flag_names"
        ],
        "title": "LoadMeasurement",
        "type": "object"
      },
...
      "PoiMeasurement": {
        "description": "The POI meter. Positive P/Q = export to the utility.",
        "properties": {
          "alarm_flag_names": {
            "items": {
              "type": "string"
            },
            "title": "Alarm Flag Names",
            "type": "array"
          },
          "alarm_flags": {
            "title": "Alarm Flags",
            "type": "integer"
          },
          "angle_deg": {
            "title": "Angle Deg",
            "type": "number"
          },
          "i_a": {
            "title": "I A",
            "type": "number"
          },
          "meter_state": {
            "title": "Meter State",
            "type": "integer"
          },
          "meter_state_name": {
            "title": "Meter State Name",
            "type": "string"
          },
          "p_kw": {
            "title": "P Kw",
            "type": "number"
          },
          "p_loss_total_kw": {
            "description": "Site losses: transformers + collector feeders.",
            "title": "P Loss Total Kw",
            "type": "number"
          },
          "pf": {
            "description": "|P|/S signed with Q (positive = exporting vars).",
            "title": "Pf",
            "type": "number"
          },
          "q_kvar": {
            "title": "Q Kvar",
            "type": "number"
          },
          "q_loss_total_kvar": {
            "title": "Q Loss Total Kvar",
            "type": "number"
          },
          "s_kva": {
            "title": "S Kva",
            "type": "number"
          },
          "v_kv": {
            "title": "V Kv",
            "type": "number"
          },
          "v_pu": {
            "title": "V Pu",
            "type": "number"
          }
        },
        "required": [
          "p_kw",
          "q_kvar",
          "s_kva",
          "pf",
          "v_kv",
          "v_pu",
          "angle_deg",
          "i_a",
          "p_loss_total_kw",
          "q_loss_total_kvar",
          "meter_state",
          "meter_state_name",
          "alarm_flags",
          "alarm_flag_names"
        ],
        "title": "PoiMeasurement",
        "type": "object"
      },
      "PoiResponse": {
        "properties": {
          "converged": {
            "title": "Converged",
            "type": "boolean"
          },
          "poi": {
            "$ref": "#/components/schemas/PoiMeasurement"
          },
          "sim_time": {
            "format": "date-time",
            "title": "Sim Time",
            "type": "string"
          },
          "step_id": {
            "title": "Step Id",
            "type": "integer"
          }
        },
        "required": [
          "step_id",
          "sim_time",
          "converged",
          "poi"
        ],
        "title": "PoiResponse",
        "type": "object"
      },
...
      "PvMeasurement": {
        "properties": {
          "alarm_flag_names": {
            "items": {
              "type": "string"
            },
            "title": "Alarm Flag Names",
            "type": "array"
          },
          "alarm_flags": {
            "title": "Alarm Flags",
            "type": "integer"
          },
          "curtailment_kw": {
            "title": "Curtailment Kw",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "inverter_state": {
            "title": "Inverter State",
            "type": "integer"
          },
          "inverter_state_name": {
            "title": "Inverter State Name",
            "type": "string"
          },
          "irradiance_wm2": {
            "title": "Irradiance Wm2",
            "type": "number"
          },
          "limit_flag_names": {
            "items": {
              "type": "string"
            },
            "title": "Limit Flag Names",
            "type": "array"
          },
          "limit_flags": {
            "title": "Limit Flags",
            "type": "integer"
          },
          "p_available_kw": {
            "title": "P Available Kw",
            "type": "number"
          },
          "p_kw": {
            "title": "P Kw",
            "type": "number"
          },
          "p_limit_active_kw": {
            "title": "P Limit Active Kw",
            "type": "number"
          },
          "pf": {
            "title": "Pf",
            "type": "number"
          },
          "q_kvar": {
            "title": "Q Kvar",
            "type": "number"
          },
          "s_kva": {
            "title": "S Kva",
            "type": "number"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "status_name": {
            "title": "Status Name",
            "type": "string"
          },
          "v_lv_kv": {
            "title": "V Lv Kv",
            "type": "number"
          },
          "v_lv_pu": {
            "title": "V Lv Pu",
            "type": "number"
          }
        },
        "required": [
          "id",
          "p_available_kw",
          "p_limit_active_kw",
          "p_kw",
          "q_kvar",
          "s_kva",
          "pf",
          "curtailment_kw",
          "irradiance_wm2",
          "status",
          "status_name",
          "limit_flags",
          "limit_flag_names",
          "v_lv_pu",
          "v_lv_kv",
          "inverter_state",
          "inverter_state_name",
          "alarm_flags",
          "alarm_flag_names"
        ],
        "title": "PvMeasurement",
        "type": "object"
      },
...
      "RunState": {
        "enum": [
          "stopped",
          "running",
          "paused"
        ],
        "title": "RunState",
        "type": "string"
      },
...
      "Snapshot": {
        "properties": {
          "bess": {
            "items": {
              "$ref": "#/components/schemas/BessMeasurement"
            },
            "title": "Bess",
            "type": "array"
          },
          "buses": {
            "items": {
              "$ref": "#/components/schemas/BusMeasurement"
            },
            "title": "Buses",
            "type": "array"
          },
          "converged": {
            "description": "False: the power flow failed this step and the values are the last good ones.",
            "title": "Converged",
            "type": "boolean"
          },
          "loads": {
            "items": {
              "$ref": "#/components/schemas/LoadMeasurement"
            },
            "title": "Loads",
            "type": "array"
          },
          "poi": {
            "$ref": "#/components/schemas/PoiMeasurement"
          },
          "pv": {
            "items": {
              "$ref": "#/components/schemas/PvMeasurement"
            },
            "title": "Pv",
            "type": "array"
          },
          "sim_time": {
            "format": "date-time",
            "title": "Sim Time",
            "type": "string"
          },
          "step_id": {
            "description": "Sim tick number; increases by 1 per step (gaps = skipped).",
            "title": "Step Id",
            "type": "integer"
          },
          "transformers": {
            "items": {
              "$ref": "#/components/schemas/TransformerMeasurement"
            },
            "title": "Transformers",
            "type": "array"
          }
        },
        "required": [
          "step_id",
          "sim_time",
          "converged",
          "poi",
          "buses",
          "transformers",
          "bess",
          "pv",
          "loads"
        ],
        "title": "Snapshot",
        "type": "object"
      },
...
      "TransformerMeasurement": {
        "description": "Positive P/Q = flowing toward the grid (LV → HV).",
        "properties": {
          "loading_pct": {
            "title": "Loading Pct",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "p_hv_kw": {
            "description": "P leaving at the HV terminal.",
            "title": "P Hv Kw",
            "type": "number"
          },
          "p_loss_kw": {
            "title": "P Loss Kw",
            "type": "number"
          },
          "p_lv_kw": {
            "description": "P entering at the LV terminal.",
            "title": "P Lv Kw",
            "type": "number"
          },
          "q_hv_kvar": {
            "title": "Q Hv Kvar",
            "type": "number"
          },
          "q_loss_kvar": {
            "title": "Q Loss Kvar",
            "type": "number"
          },
          "q_lv_kvar": {
            "title": "Q Lv Kvar",
            "type": "number"
          }
        },
        "required": [
          "name",
          "p_hv_kw",
          "q_hv_kvar",
          "p_lv_kw",
          "q_lv_kvar",
          "p_loss_kw",
          "q_loss_kvar",
          "loading_pct"
        ],
        "title": "TransformerMeasurement",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Grid-connected microgrid power flow simulator (pandapower, balanced positive sequence,\nquasi-static). The EMS writes setpoints; every step the simulator applies the asset limits,\nsolves the network and publishes a measurement snapshot.\n\n**Sign conventions**\n- BESS / PV P and Q: generator convention. P > 0 injects (BESS discharging, PV producing),\n  P < 0 is a charging BESS. Q > 0 injects vars.\n- Load P and Q: P > 0 consumes, Q > 0 consumes vars (lagging).\n- POI: P > 0 exports to the utility, P < 0 imports. The same for Q.\n- Transformers: P/Q > 0 flows toward the grid (LV → HV).\n- pf = |P|/S, signed with Q in the element's own convention.\n\n**Units:** kW, kvar, kVA, kWh, kV (line-to-line), pu where noted, s.\n\n**Points:** `<asset_type>.<asset_id>.<point>`, listed at GET /api/points.\n",
    "title": "powerflow",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/measurements/history": {
      "get": {
        "operationId": "history_api_measurements_history_get",
        "parameters": [
          {
            "description": "Sim time ≥ from.",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sim time ≥ from.",
              "title": "From"
            }
          },
          {
            "description": "Sim time ≤ to.",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sim time ≤ to.",
              "title": "To"
            }
          },
          {
            "description": "Comma-separated measurement point names (e.g. `poi.meter.p_kw,bess.bess1.soc_pct`). Default: every measurement point.",
            "in": "query",
            "name": "fields",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated measurement point names (e.g. `poi.meter.p_kw,bess.bess1.soc_pct`). Default: every measurement point.",
              "title": "Fields"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/HistoryFormat",
              "default": "json"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown field."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Field is not a measurement point."
          }
        },
        "summary": "Snapshots from the in-memory ring buffer, as rows of point values",
        "tags": [
          "measurements"
        ]
      }
    },
    "/api/measurements/latest": {
      "get": {
        "description": "POI, every bus, every transformer, every asset (commanded vs actual), losses and the convergence flag. `step_id` increases by one per sim tick, so a poller can spot missed steps and backfill from /measurements/history.",
        "operationId": "latest_api_measurements_latest_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No step has run yet."
          }
        },
        "summary": "The full latest snapshot",
        "tags": [
          "measurements"
        ]
      }
    },
    "/api/measurements/poi": {
      "get": {
        "operationId": "poi_api_measurements_poi_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No step has run yet."
          }
        },
        "summary": "The POI meter",
        "tags": [
          "measurements"
        ]
      }
    },
...
    "/api/sim/pause": {
      "post": {
        "operationId": "pause_api_sim_pause_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngineStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed in the current state."
          }
        },
        "summary": "Pause",
        "tags": [
          "simulation"
        ]
      }
    },
...
    "/api/sim/step": {
      "post": {
        "description": "Only with `simulation.test_mode: true` and while not running; otherwise 409. Returns the last step's snapshot.",
        "operationId": "step_api_sim_step_post",
        "parameters": [
          {
            "in": "query",
            "name": "count",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 1000000,
              "minimum": 1,
              "title": "Count",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed in the current state."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Advance N steps at once (test mode only)",
        "tags": [
          "simulation"
        ]
      }
    },
...
  }
}